fix(types): 16 mypy errors, all of them mine, before CI ever reached the step
`make type` is the fourth step of CI's lint job and had never run — the job died
at `lint-markdown` ahead of it. With that unblocked it would have gone straight
to 16 errors. I blamed every offending line: 16 from commits on this branch, 0
pre-existing.

I could not run mypy the way CI does. It dies inside a numpy stub
(`numpy/__init__.pyi:737: Type statement is only supported in Python 3.12 and
greater`) before reading any of our source, because the config pins
python_version 3.10 and this venv is 3.12 with newer numpy stubs. So this was
checked with `--python-version 3.12` — a diagnostic override, no config change.
It reads OUR code, which is the point.

Two of the sixteen were real:

* `_affinity_boost` declared `-> float` and returned `weight * min(saturated,
  cap)` unconverted. Both operands come out of the ranking config as `Any`, so
  the declared return type was never enforced — the one signature in the ranking
  path where a config typo could have flowed onward as a non-float.

* `test_summary_poison_guard` asserted
  `_reject_if_prompt_examples_leaked(...) is None` — twice. That function is
  declared `-> None`, so the comparison is true no matter what it does. Another
  assertion that could not fail. The real contract is "does not raise", which is
  what a bare call states, and the meaningful assertion (no SUMMARY POISONED in
  the log) was already sitting underneath it.

The rest were test typing: three `None`s passed where `CatalogEpisodeRow`
declares `str` (production always passes a path — `has_gi`/`has_kg` carry
existence, so `""` is the honest stand-in), `Any` leaking through declared
return types, and dict literals inferred as `dict[str, object]`.

Three of them were the `lst.append(x) or y` idiom inside monkeypatch lambdas —
sequencing two statements in an expression by leaning on `append` returning
None. Replaced with named recorder functions that do the same job and say so.

  mypy --config-file pyproject.toml --python-version 3.12 .
    Success: no issues found in 1724 source files   (was: 16 errors in 10 files)
  flake8 / black / isort   clean
  the 9 test files touched  228 passed, 2 xfailed

+++ src/podcast_scraper/server/app_discover_view.py
@@ -83,7 +83,9 @@ def _affinity_boost(
         return 0.0
     # 1 - 0.5**n : one match gives half the cap, two gives three quarters, and it never exceeds it.
     saturated = 1.0 - 0.5**contribution
-    return weight * min(saturated, cap)
+    # float(): weight/cap come out of the ranking config as Any, so returning the product
+    # unconverted made the declared -> float unenforced.
+    return float(weight * min(saturated, cap))
 
 
 def _feed_significance_means(
